% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directional.R
\name{directional}
\alias{directional}
\title{Perform Directional Pleiotropy Test}
\usage{
directional(u)
}
\arguments{
\item{u}{Numeric vector of direct genetic effect of the outcome.}
}
\value{
A list containing:
\item{mean}{The mean of the direct genetic effect of outcome \code{u} estimated from the t-test.}
\item{p_value}{The p-value from the t-test.}
}
\description{
This function performs a one-sample t-test to assess whether the mean of the residuals (\code{u}) is significantly different from zero. This test helps in evaluating the presence of directional pleiotropy by checking if the residuals deviate significantly from zero.
}
\details{
The function performs a t-test where the null hypothesis is that the mean of \code{u} is zero. The result includes the estimated mean of \code{u} and the p-value from the t-test.
}
