% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRLOVA.R
\name{mr_lova}
\alias{mr_lova}
\title{Mendelian Randomization Analysis}
\usage{
mr_lova(
  betaY,
  betaX,
  betaYse,
  betaXse,
  ny,
  gwas_p = 0.05,
  gwas_p2 = 5e-08,
  permutn = 0,
  log_file = NULL
)
}
\arguments{
\item{betaY}{beta of the outcome, recommended to be in standardized scale.}

\item{betaX}{beta of the exposure, recommended to be in standardized scale.}

\item{betaYse}{standard error of the outcome, recommended to be in standardized scale.}

\item{betaXse}{standard error of the exposure, recommended to be in standardized scale.}

\item{ny}{sample size of outcome GWAS.}

\item{gwas_p}{p-value threshold for the association between variants and the latent phenotype (the direct genetic effect). Default: 5e-2.}

\item{gwas_p2}{p-value threshold for exposure GWAS used in instrument variable selection. Default: 5e-8}

\item{permutn}{number of permutations, Default: 0.}

\item{log_file}{file name for saving the iterations to check convergence. Default: NULL.}
}
\value{
A list that contains
\item{CausEst}{Estimate of causal effect.}
\item{CausEstSE}{Standard error of causal effect estimate.}
\item{CausEstP}{P-value from the z test for the causal effect.}
\item{IVs}{p-values for the direct causal effect of all Instrumental Variants (IV) on the outcome.}
\item{Valid}{Index of valid  instrumental variants.}
\item{corrected_p-value}{p-value from permutation tests.}
}
\description{
This function conducts Mendelian randomization analysis based on the latent phenotype of the outcome, explicitly excluding horizontal pleiotropy variants. It iteratively refines the causal relationship through an Expectation-Maximization (EM) algorithm. MRLOVA uses GWAS summary statistics as inputs to estimate the causal effects of one trait on another. We recommend using summary statistics on a standardized scale.
}
\details{
None.
}
\keyword{Mendelian}
\keyword{randomization.}
